merge roi check lai xem thu

+++ src/java/Controller/AddHouseServlet.java
@@ -9,14 +9,14 @@ import Dao.HouseDAO;
 import Model.House;
 import Model.Location;
 import Model.Menu;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 /**
  *
@@ -88,7 +88,7 @@ public class AddHouseServlet extends HttpServlet {
         }
         String housename = request.getParameter("housename");
         String review = request.getParameter("review");
-        float price = Float.parseFloat(request.getParameter("houseprice"));
+        double price = Double.parseDouble(request.getParameter("houseprice"));
         int status = 1;
         String address = request.getParameter("address");
         String description = request.getParameter("description");
